Rename peopleImport to simply "importer"
