+++ src/store/index.js
@@ -13,7 +13,7 @@ import locations from './locations';
 import org from './org';
 import participants from './participants';
 import people from './people';
-import peopleImport from './peopleImport';
+import importer from './importer';
 import personTags from './personTags';
 import queries from './queries';
 import search from './search';
@@ -33,7 +33,7 @@ export const appReducer = combineReducers({
     org,
     participants,
     people,
-    peopleImport,
+    importer,
     personTags,
     queries,
     search,
